+++ src/Pages/FolderPage.js
@@ -1,7 +1,3 @@
-
-
-
-
 import useAsync from '../Hooks/useAsync'
 import { getUser } from '../api'
 import AddLink from '../components/AddLink/AddLink'
@@ -10,7 +6,6 @@ import Footer from '../components/Footer/Footer'
 import Main from '../components/Main/Main'
 import Nav from '../components/Nav/Nav'
 import SearchBar from '../components/SearchBar/SearchBar'
-
 
 function FolderPage() {
   const [data, isLoading, hasError] = useAsync(() => getUser(1))
