+++ src/components/Nav/Profile.jsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import * as S from './ProfileStyle';
+import React from 'react'
+import * as S from './ProfileStyle'
 
 function Profile({ email, profileImage }) {
   return (
@@ -7,7 +7,7 @@ function Profile({ email, profileImage }) {
       <S.ProfileImg src={profileImage} alt="프로필 사진" />
       <S.ProfileEmail className="flex-center">{email}</S.ProfileEmail>
     </S.Profile>
-  );
+  )
 }
 
-export default Profile;
+export default Profile
